Refactors as /flow-events route
