+++ lib/middleware/airtable.js
@@ -4,16 +4,37 @@ const logger = require('heroku-logger');
 const superagent = require('superagent');
 const { assign, startCase } = require('lodash');
 
-const airtable = require('../../services/airtable');
-const { findOrCreateAirtableRecord } = require('../../helpers');
+const airtable = require('../services/airtable');
+
+/**
+ * @param {String} tableName
+ * @param {Object} fields
+ * @return {Boolean}
+ */
+const findOrCreateAirtableRecord = async (tableName, fields) => {
+  let record = await airtable.getRecordByUuid(tableName, fields.Uuid);
+
+  if (record) {
+    logger.debug('Found Airtable record', { tableName, record });
+
+    return record;
+  }
+
+  record = await airtable.createRecord(tableName, fields);
+
+  logger.debug('Created Airtable record', { tableName, record });
+
+  return record;
+};
+
 
 module.exports = function createRecord() {
   return async (req, res, next) => {
     try {
-      const flowEventTableName = startCase(req.query.table);
+      const flowEventTableName = startCase(req.query.airtable);
 
       if (!flowEventTableName) {
-        return res.status(422).send('Missing table query parameter');
+        return next();
       }
 
       const contactRecord = await findOrCreateAirtableRecord('Contacts', req.contact);
