--- lib/middleware/export/formatData.js
@@ -1,68 +1,0 @@
-'use strict';
-
-const lodash = require('lodash');
-const logger = require('heroku-logger');
-
-/**
- * @param {String} fieldName
- * @param {String} fieldValue
- * @return {String}
- */
-const formatPlainTextField = (fieldName, fieldValue) => {
-  let value = fieldValue;
-
-  if (fieldName === 'created_on') {
-    // Get the leading YYYY-MM-DD value.
-    value = value.substring(0, 10);
-  }
-
-  return `${lodash.startCase(fieldName)}: ${value}`;
-}
-
-module.exports = function formatData() {
-  return async (req, res, next) => {
-    try {
-      const plainTextResult = [];
-
-      const contactFields = [
-        'name',
-        'phone',
-        'url',
-        'created_on',
-        'groups'
-      ];
-
-      contactFields.forEach(fieldName => plainTextResult.push(formatPlainTextField(
-        fieldName,
-        req.data[fieldName],
-      )));
-
-      plainTextResult.push(' ');
-
-      plainTextResult.push(formatPlainTextField('flow', req.data.flow.name));
-
-      Object.keys(req.data.results).forEach((fieldName) => {
-        plainTextResult.push(formatPlainTextField(
-          `${fieldName} category`,
-          req.data.results[fieldName].category,
-        ));
-        plainTextResult.push(formatPlainTextField(
-          `${fieldName} value`,
-          req.data.results[fieldName].value,
-        ));
-      });
-
-      plainTextResult.push(' ');
-
-      req.data.fields ? Object.keys(req.data.fields).sort().forEach((fieldName) => {
-        plainTextResult.push(formatPlainTextField(fieldName, req.data.fields[fieldName]));
-      }) : null;
-
-      req.data = { text: plainTextResult.join('\n') };
-
-      return next();
-    } catch (error) {
-      return next(error);
-    }
-  }
-};
