+++ lib/middleware/parseFlowEvent.js
@@ -25,7 +25,10 @@ module.exports = function parseFlowEvent() {
 
       req.contact = formatters.formatTextItContact(textItRes ? textItRes : contact);
 
-      req.text = formatters.getPlainTextFromReq(req);
+      // Save the payload to send to Zapier and as the response to this API request.
+      req.data = req.contact;
+      req.Flow = req.flow.name;
+      assign(req.data, req.results);
 
       return next();
     } catch (error) {
