+++ lib/middleware/sendResponse.js
@@ -3,15 +3,21 @@
 const { assign } = require('lodash');
 const logger = require('heroku-logger');
 
+const { getPlainTextFromReq } = require('../formatters');
+
 module.exports = function sendResponse() {
-  return (req, res) => {
-    const data = req.data;
+  return (req, res, next) => {
+    try {
+      const data = req.data;
 
-    // Save plain text to use in transactional emails from TextIt.
-    data.Text = req.text;
+      logger.debug('Sending data', { data });
 
-    logger.debug('Sending data', { data} );
-
-    res.send(data);
+      return res.send({
+        data,
+        text: getPlainTextFromReq(req),
+      });
+    } catch (error) {
+      return next(error);
+    }
   };
 };
