+++ lib/middleware/zapier.js
@@ -4,19 +4,12 @@ const { assign } = require('lodash');
 const logger = require('heroku-logger');
 const superagent = require('superagent');
 
-const zapier = require('../../services/zapier');
+const zapier = require('../services/zapier');
 
 module.exports = function postWebhook() {
   return async (req, res, next) => {
     try {
-      req.data = {
-        'Contact': req.contact,
-        'Flow': assign(req.flow, req.results),
-      };
-
-      //assign(req.data.flow, req.results);
-
-      if (req.query.test) {
+      if (!req.query.zapier) {
         logger.debug('Skipping send to Zapier');
 
         return next();
