+++ routes.js
@@ -2,12 +2,11 @@
 
 const logger = require('heroku-logger');
 
+const airtableMiddleware = require('./lib/middleware/zapier');
+const zapierMiddleware = require('./lib/middleware/airtable');
 const authenticateMiddleware = require('./lib/middleware/authenticate');
 const sendResponseMiddleware = require('./lib/middleware/sendResponse');
-const formatDataMiddleware = require('./lib/middleware/export/formatData');
 const parseFlowEventMiddleware = require('./lib/middleware/parseFlowEvent');
-const postZapierWebhookMiddleware = require('./lib/middleware/zapier/postWebhook');
-const createAirtableRecordMiddleware = require('./lib/middleware/airtable/createRecord');
 
 /**
  * API routes.
@@ -17,19 +16,10 @@ module.exports = (app) => {
 
   app.get('/', (req, res) => res.send('OK'));
 
-  app.post('/api/v1/airtable',
+  app.post('/api/v1/flow-events',
     parseFlowEventMiddleware(),
-    createAirtableRecordMiddleware(),
-    sendResponseMiddleware());
-
-  app.post('/api/v1/export',
-    parseFlowEventMiddleware(),
-    formatDataMiddleware(),
-    sendResponseMiddleware());
-
-  app.post('/api/v1/zapier/:id1/:id2',
-    parseFlowEventMiddleware(),
-    postZapierWebhookMiddleware(),
+    zapierMiddleware(),
+    airtableMiddleware(),
     sendResponseMiddleware());
 
   // Error handler
